requirements.txt
alembic==1.10.2
click==8.1.3
Flask==2.2.3
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
pyaes==1.6.1
Pyrogram==2.0.102
PySocks==1.7.1
PyYAML==6.0
SQLAlchemy==2.0.6
typing_extensions==4.5.0
uvloop==0.17.0
Werkzeug==2.2.3
